Combine EncodeHeaderHandler and PrefillHeaderHandler into a single DisaggHeadersHandler (#758)
* combine EncodeHeaderHandler PrefillHeaderHandler to DisaggHeadersHandler


* rename disagg_prerequest.go to disagg_headers_handler.go


---------
