+++ deploy/config/pd-epp-config.yaml
@@ -4,7 +4,7 @@ kind: EndpointPickerConfig
 featureGates:
 - prepareDataPlugins
 plugins:
-- type: prefill-header-handler
+- type: disagg-headers-handler
 - type: prefix-cache-scorer
   parameters:
     maxPrefixBlocksToMatch: 256
