+++ docs/architecture.md
@@ -199,13 +199,14 @@ If the configuration is in a file, the EPP command line argument `--configFile` 
 
 This section describes how to setup the various plugins available with the llm-d-inference-scheduler
 
-#### PrefillHeader
+#### DisaggHeadersHandler
 
-Sets a header for use in disaggregated prefill/decode
+Sets headers for use in disaggregated prefill/decode and encode/prefill/decode
 
-- **Type**: `prefill-header-handler`
+- **Type**: `disagg-headers-handler`
 - **Parameters**:
   - `prefillProfile`: specifies the name of the profile used for the prefill scheduling. Only needed if the prefill profile is not named `prefill`.
+  - `encodeProfile`: specifies the name of the profile used for the encode scheduling. Only needed if the encode profile is not named `encode`.
 
 ---
 
@@ -583,7 +584,7 @@ The following is an example of what a configuration for disaggregated Prefill/De
 apiVersion: inference.networking.x-k8s.io/v1alpha1
 kind: EndpointPickerConfig
 plugins:
-- type: prefill-header-handler
+- type: disagg-headers-handler
 - type: prefix-cache-scorer
   parameters:
     maxPrefixBlocksToMatch: 256
