+++ docs/disaggregation.md
@@ -277,7 +277,7 @@ plugins:
       maxPrefixBlocksToMatch: 256
       lruCapacityPerServer: 31250
   - type: max-score-picker
-  - type: prefill-header-handler
+  - type: disagg-headers-handler
   - type: prefix-based-pd-decider
     parameters:
       nonCachedTokens: 8
@@ -336,8 +336,7 @@ plugins:
       maxPrefixBlocksToMatch: 256
       lruCapacityPerServer: 31250
   - type: max-score-picker
-  - type: encode-header-handler
-  - type: prefill-header-handler
+  - type: disagg-headers-handler
   - type: always-disagg-multimodal-decider
   - type: prefix-based-pd-decider
     parameters:
